1) Support of 'delete obj[key|index]' and 'delete obj' operators, handling is similar to JS. 2) bultin constant '_next' was renamed to '!next' to disable its conflicts with user variables 3) local variable declarations in if() and while() statements. So it is possible now to define something like this:    if( var t = ... )    {       t.foo();    }
4) fix of try {} finally {} handling (no catch() section)
 
5) include ... 
    a ) only relative paths are stored in bytecode, used to store absolute paths at compilation.
    
6) New macro variables: __FILE__, __LINE__, and __TRACE__

7) New, support of binary literals: 0b00100010 == 0x22;

8) Support of multi-return-values in native API, set_nth_retval()

9) New Bytes.name property.

10) New Color.tint() method.

11) operator == is made more JS compatible when obejact of different types are involved.

12) Math.atan2() function added

13) String.$() stringizer function.

14) XMLScannerscan() returns WORD and SPACE tokens - more granular parsing.

15) TIScript is using Rob Pike RE package for regular expressions. More reliable and faster.

+++ trunk/tool/tl_pool.h
@@ -233,7 +233,7 @@ namespace tool
   inline bool
     pool<c_element,c_traits>::exists ( const c_element& the_key )
   {
-    return ( get_index ( the_key, false ) != (size_t)-1 );
+    return ( get_index ( the_key, false ) != uint(-1) );
   }
 
   template <typename c_element, typename c_traits>
